updates to dashboard and hcare folder structure
